The final version

+++ application.py
@@ -1,4 +1,6 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
+#!/usr/bin/env python3
+from flask import Flask, render_template, request, redirect, jsonify
+from flask import url_for, flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Category, Item, User
@@ -17,14 +19,17 @@ import requests
 # or do the list ob object idea
 app = Flask(__name__)
 
-engine = create_engine('sqlite:///itemcatalog.db',  poolclass=SingletonThreadPool)
+engine = create_engine(
+    'sqlite:///itemcatalog.db', poolclass=SingletonThreadPool)
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
+CLIENT_ID = json.loads(open(
+    'client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Item catalog Application"
+
 
 # Create anti-forgery state token
 @app.route('/login')
@@ -33,6 +38,7 @@ def showLogin():
                     for x in xrange(32))
     login_session['state'] = state
     return render_template('login.html', STATE=state)
+
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -79,15 +85,15 @@ def gconnect():
     if result['issued_to'] != CLIENT_ID:
         response = make_response(
             json.dumps("Token's client ID does not match app's."), 401)
-        print ("Token's client ID does not match app's.")
+        print("Token's client ID does not match app's.")
         response.headers['Content-Type'] = 'application/json'
         return response
 
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
-                                 200)
+        response = make_response(
+            json.dumps('Current user is already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -117,8 +123,9 @@ def gconnect():
     output += login_session['username']
     output += '!</h1>'
     flash("you are now logged in as %s" % login_session['username'])
-    print ("done!")
+    print("done!")
     return output
+
 
 # User Helper Functions
 def createUser(login_session):
@@ -139,8 +146,9 @@ def getUserID(email):
     try:
         user = session.query(User).filter_by(email=email).one()
         return user.id
-    except:
+    except Exception:
         return None
+
 
 # DISCONNECT - Revoke a current user's token and reset their login_session
 @app.route('/gdisconnect')
@@ -163,19 +171,22 @@ def gdisconnect():
         del login_session['username']
         del login_session['email']
         del login_session['user_id']
-        flash("you loggedout successfully")
-        return redirect('/')#redirect to show catalog then do the other steps to add delete edit
+        flash("you logged out successfully")
+        return redirect('/')
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
 
-#JSON API for catalog categories
+
+# JSON API for catalog categories
 @app.route('/<int:category_id>/JSON')
 def categoryItemsJSON(category_id):
     items = session.query(Item).filter_by(
         category_id=category_id).all()
     return jsonify(Category=[category.serialize for category in items])
+
 
 # JSON API for a category items
 @app.route('/<int:category_id>/<int:item_id>/JSON')
@@ -183,41 +194,48 @@ def categoryItemJSON(category_id, item_id):
     item = session.query(Item).filter_by(id=item_id).one()
     return jsonify(Item=item.serialize)
 
+
 # Show all categories
 @app.route('/')
 @app.route('/catalog/')
 def showCategories():
     categories = session.query(Category)
     Items = session.query(Item).all()
-    counter=0
-    length= len(Items)
-    print(length)
-    latestItems =[]
+    counter = 0
+    length = len(Items)
+    latestItems = []
     latestItemCat = []
-    if length<=10:
-        while counter<length:
-            print(counter)
-            print(Items[counter].title)
+    if length <= 10:
+        while counter < length:
             latestItems.append(Items[counter])
             D = Items[counter]
-            cat = session.query(Category).filter_by(id = D.category_id).one()
+            cat = session.query(Category).filter_by(id=D.category_id).one()
             latestItemCat.append(cat)
-            counter+=1
-        return render_template('catalog.html', categories = categories,theList = zip(latestItems,latestItemCat),pictureURL=login_session['picture'])
+            counter += 1
+        if 'username' not in login_session:
+            return render_template('catalog.html', categories=categories,
+                                   theList=zip(latestItems, latestItemCat))
+        else:
+            return render_template('catalog.html', categories=categories,
+                                   pictureURL=login_session['picture'],
+                                   theList=zip(latestItems, latestItemCat))
     else:
-        while length-counter !=10:
-            counter+=1
-        
+        while length - counter != 10:
+            counter += 1
         while length-counter != 0:
             latestItems.append(Items[counter])
             D = Items[counter]
-            cat = session.query(Category).filter_by(id = D.category_id).one()
+            cat = session.query(Category).filter_by(id=D.category_id).one()
             latestItemCat.append(cat)
-            counter+=1
+            counter += 1
         if 'username' not in login_session:
-            return render_template('catalog.html', categories = categories,theList = zip(latestItems,latestItemCat))
+            return render_template('catalog.html', categories=categories,
+                                   theList=zip(latestItems, latestItemCat))
         else:
-            return render_template('catalog.html', categories = categories,pictureURL=login_session['picture'],theList = zip(latestItems,latestItemCat))
+            return render_template('catalog.html', categories=categories,
+                                   pictureURL=login_session['picture'],
+                                   theList=zip(latestItems, latestItemCat))
+
 
 # Show all category items
 @app.route('/catalog/<int:category_id>/')
@@ -226,60 +244,90 @@ def categoryDetails(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category.id).all()
     if 'username' not in login_session:
-        return render_template('publicCategoryDetails.html', items=items, category=category)
-    return render_template('categoryDetails.html', items=items, category=category, category_id = category_id,pictureURL=login_session['picture'])
+        return render_template('publicCategoryDetails.html',
+                               items=items, category=category)
+    return render_template('categoryDetails.html', items=items,
+                           category=category, category_id=category_id,
+                           pictureURL=login_session['picture'])
+
 
 # show item description
 @app.route('/catalog/<int:category_id>/<int:item_id>')
 @app.route('/catalog/<int:category_id>/<int:item_id>/details')
-def itemDetails(category_id,item_id):
+def itemDetails(category_id, item_id):
     category = session.query(Category).filter_by(id=category_id).one()
-    item = session.query(Item).filter_by(category_id=category_id,id = item_id).one()
+    item = session.query(Item).filter_by(category_id=category_id,
+                                         id=item_id).one()
     if 'username' not in login_session:
-        return render_template('publicItemDetails.html', item=item, category=category)
-    return render_template('itemDetails.html', item=item, category=category,pictureURL=login_session['picture'])
+        return render_template('publicItemDetails.html', item=item,
+                               category=category)
+    return render_template('itemDetails.html', item=item, category=category,
+                           pictureURL=login_session['picture'])
 
-# Add new category item --- for main add item loop throw the Category and save it to variable 
+
+# Add new category item
 @app.route('/catalog/<int:category_id>/details/new/', methods=['GET', 'POST'])
 def newCategoryItem(category_id):
     if 'username' not in login_session:
         return redirect('/login')
     category = session.query(Category).filter_by(id=category_id).one()
     if request.method == 'POST':
-        newItem = Item(title=request.form['title'], description=request.form['description'], category_id=category_id)
+        newItem = Item(title=request.form['title'],
+                       description=request.form['description'],
+                       category_id=category_id,
+                       user_id=login_session['user_id'])
         session.add(newItem)
         session.commit()
         flash('New category Item %s Successfully Created' % (newItem.title))
-        return redirect(url_for('categoryDetails', category_id=category_id,pictureURL=login_session['picture']))
+        return redirect(url_for('categoryDetails', category_id=category_id,
+                                pictureURL=login_session['picture']))
     else:
-        return render_template('newCategoryItem.html', category=category,pictureURL=login_session['picture'])
+        return render_template('newCategoryItem.html', category=category,
+                               pictureURL=login_session['picture'])
 
-#for main add
+
+# for main add
 @app.route('/catalog/new', methods=['GET', 'POST'])
 def newItem():
     if 'username' not in login_session:
         return redirect('/login')
     if request.method == 'POST':
         if request.form['title']:
-            category = session.query(Category).filter_by(name= request.form['category']).one()
-            newItem = Item(title=request.form['title'], description=request.form['description'], category_id=category.id)
+            category = session.query(Category).filter_by(
+                name=request.form['category']).one()
+            newItem = Item(title=request.form['title'],
+                           description=request.form['description'],
+                           category_id=category.id,
+                           user_id=login_session['user_id'])
             session.add(newItem)
             session.commit()
-            flash('New category Item %s Successfully Created' % (newItem.title))
-            return redirect(url_for('categoryDetails', category_id=category.id,pictureURL=login_session['picture']))
+            flash(
+                'New category Item %s Successfully Created' % (newItem.title))
+            return redirect(url_for('categoryDetails',
+                                    category_id=category.id,
+                                    pictureURL=login_session['picture']))
     else:
-        return render_template('newItemForm.html',pictureURL=login_session['picture'])
+        return render_template('newItemForm.html',
+                               pictureURL=login_session['picture'])
 
-# Edit item 
-@app.route('/catalog/<int:category_id>/<int:item_id>/edit/', methods=['GET', 'POST'])
-def editItem(category_id,item_id):
+
+# Edit item
+@app.route('/catalog/<int:category_id>/<int:item_id>/edit/',
+           methods=['GET', 'POST'])
+def editItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
     category = session.query(Category).filter_by(id=category_id).one()
-    editedItem = session.query(Item).filter_by(category_id=category_id,id = item_id).one()
+    editedItem = session.query(Item).filter_by(category_id=category_id,
+                                               id=item_id).one()
+    if login_session['user_id'] != editedItem.user_id:
+        flash("You are not authorized to edit this item!")
+        return redirect(url_for('itemDetails', category_id=category_id,
+                                item_id=item_id,
+                                pictureURL=login_session['picture']))
     if request.method == 'POST':
         Cname = request.form['category']
-        newC = session.query(Category).filter_by(name = Cname).one()
+        newC = session.query(Category).filter_by(name=Cname).one()
         newID = newC.id
         if request.form['title']:
             editedItem.title = request.form['title']
@@ -290,27 +338,42 @@ def editItem(category_id,item_id):
             session.add(newItem)
             session.commit()
             flash('%s Successfully Edited' % editedItem.title)
-            return redirect(url_for('itemDetails',category_id =newID , item_id = newItem.id,pictureURL=login_session['picture']))
+            return redirect(url_for('itemDetails', category_id=newID,
+                                    item_id=newItem.id,
+                                    pictureURL=login_session['picture']))
     else:
-        return render_template('editItem.html', item=editedItem, category = category,pictureURL=login_session['picture'])
+        return render_template('editItem.html', item=editedItem,
+                               category=category,
+                               pictureURL=login_session['picture'])
+
 
 # Delete category item
-@app.route('/catalog/<int:category_id>/<int:item_id>/delete/', methods=['GET', 'POST'])
-def deleteItem(category_id,item_id):
+@app.route('/catalog/<int:category_id>/<int:item_id>/delete/',
+           methods=['GET', 'POST'])
+def deleteItem(category_id, item_id):
     if 'username' not in login_session:
         return redirect('/login')
     category = session.query(Category).filter_by(id=category_id).one()
-    deletedItem = session.query(Item).filter_by(category_id=category.id,id = item_id).one()
+    deletedItem = session.query(Item).filter_by(category_id=category.id,
+                                                id=item_id).one()
+    if login_session['user_id'] != deletedItem.user_id:
+        flash("You are not authorized to delete this item!")
+        return redirect(url_for('itemDetails', category_id=category_id,
+                                item_id=item_id,
+                                pictureURL=login_session['picture']))
     if request.method == 'POST':
         session.delete(deletedItem)
         session.commit()
         flash('Item %s Successfully deleted' % (deletedItem.title))
-        return redirect(url_for('categoryDetails', category_id=category_id,pictureURL=login_session['picture']))
+        return redirect(url_for('categoryDetails', category_id=category_id,
+                                pictureURL=login_session['picture']))
     else:
-        return render_template('deleteItem.html', category=category,item = deletedItem,pictureURL=login_session['picture'])
+        return render_template('deleteItem.html', category=category,
+                               item=deletedItem,
+                               pictureURL=login_session['picture'])
 
 
 if __name__ == '__main__':
     app.secret_key = 'super_secret_key'
     app.debug = True
-    app.run(host='0.0.0.0', port=5000,threaded =False)
+    app.run(host='0.0.0.0', port=5000, threaded=False)
